+++ webcomponents/inline-editor/src/components/actions/color-actions/color-actions.scss
