+++ webcomponents/inline-editor/src/components/actions/color-actions/readme.md
@@ -21,7 +21,7 @@
 
 ### Used by
 
-- [deckgo-inline-editor](../inline-editor)
+- [deckgo-inline-editor](../../inline-editor)
 
 ### Depends on
