+++ webcomponents/inline-editor/src/components/actions/link-actions/readme.md
@@ -22,7 +22,7 @@
 
 ### Used by
 
-- [deckgo-inline-editor](../inline-editor)
+- [deckgo-inline-editor](../../inline-editor)
 
 ### Graph
