+++ webcomponents/inline-editor/src/components/components/action-button/action-button.scss
