+++ webcomponents/inline-editor/src/components/components/action-button/readme.md
@@ -20,8 +20,8 @@
 
 ### Used by
 
-- [deckgo-ie-image-actions](../image-actions)
-- [deckgo-inline-editor](../inline-editor)
+- [deckgo-ie-image-actions](../../actions/image-actions)
+- [deckgo-inline-editor](../../inline-editor)
 
 ### Graph
