+++ webcomponents/inline-editor/src/components/components/action-image/action-image.scss
