+++ webcomponents/inline-editor/src/components/components/separator/separator.scss
